+++ CHANGELOG.md
@@ -37,6 +37,7 @@ v5.3.0
 - Enable ARM macOS runner, fixes for SCOTCH allocation and PETSc detection on macOS (!1462)
 - Add FieldConvert module and filter to project velocity into body-fitted coordinate system (!1467)
 - Fix typos in Vmath and VDmath (!1480)
+- Fix minor typo and removed unused functions in LibUtilities/TimeIntegration (!1476) 
 
 **Python**
 - Add wrappers for Interpreter and Equation classes (!1329)